Progress: Persist state across powercycles!
One problem remains (worked around, but it's not perfect:

When `fido_authenticator::State` is created, `crypto_service::Client`
can not yet be called (at least not with the RTFM "system call";
we could try to pass in `crypto_service::Service` with its "direct"
`Syscall` implementation, but I assume this will be bad later on
when we TrustZone-ify.

For `state.identity`, which contains AAGUID and attn key, there is
anyway no solution yet - injecting a private key directly without
existing wrapping key probably should not be part of the API anyway.

For `state.persistent`, we hack in a `initialised` flag and hope
for the best (since we want to avoid infesting *all* its methods
with the generic parameters).