Remove associated data in wrapping of keys
Trussed itself already ignored this associated data (https://github.com/trussed-dev/trussed/pull/108),
and the unwrapping was already performed with no associated data.
Not removing it would lead to breakage once (https://github.com/trussed-dev/trussed/pull/108) is merged.
Adding the AD to the unwrapping step would break compatibility with currently registerd credentials.

Security: This is not an issue because the credentials stored locally are encrypted with
the proper app id as associated data which is checked when the credential is decrypted.