Extract PIN protocol implementation
This patch extracts the implementation of PIN protocol one from the
ctap2 and state modules into the ctap2.pin module.  This makes it
easier to add support for PIN protocol two in the future.

Functionally, the extracted implementation should be mostly identical to
the old implementation, expect for some smaller changes for improved
compliance with the specification (error codes, explicit token resets).