ctap2: Explicitly specify PIN protocol version
This patch removes the assumption that we only support PIN protocol 1
from the CTAP2 implementation.  Instead, a list of supported PIN
protocols is provided by the pin_protocols function and
parse_pin_protocol can be used to validate the PIN protocol selected by
the platform.  The parsed PIN protocol version must then be passed to
the pin_protcol function to access the correct implementation.