Optimize key and PIN token generation
This patch generates the key agreement key directly after the reset to
improve the execution time of the next requests.  It also refactors the
PIN token generation and storage so that the PIN token is only generated
when needed, and only for the PIN protocol that is currently in use.