get_assertion: Skip attStmt unless requested
For makeCredential, a missing attestation format preference list
means that we should use the default format (packed).  For getAssertion,
it means that we should skip the attestation statement entirely.
Previously, we implemented the makeCredential algorithm for both cases.
This caused an incompatibility with firefox because it fails on
unexpected fields in the response (in this case, the attestation
statement).  This patch fixes this issue and applies the correct default
for getAssertion requests.

Fixes: https://github.com/Nitrokey/fido-authenticator/issues/98