Check credential limit in get_creds_metadata
We have two limits for the credential count:  a fixed limit determined
by the configuration and an estimated limit based on the remaining
filesystem size.  When creating a new credential, we check both.  But
previously we only returned the estimated limit from get_creds_metadata.
This patch adds the fixed limit to get_creds_metadata.